Add Ebt Balance Flow
