+++ example/app.json
@@ -1,6 +1,6 @@
 {
   "expo": {
-    "name": "Benny Apply Example",
+    "name": "Benny SDK Example",
     "slug": "bennyapi",
     "version": "1.0.0",
     "orientation": "portrait",
