+++ example/package.json
@@ -8,7 +8,7 @@
     "ios": "expo start --ios"
   },
   "dependencies": {
-    "expo": "^50.0.7",
+    "expo": "^50.0.11",
     "expo-status-bar": "^1.11.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
