+++ package.json
@@ -59,22 +59,22 @@
     "@react-native/eslint-config": "^0.73.2",
     "@release-it/conventional-changelog": "8.0.1",
     "@types/jest": "^29.5.12",
-    "@types/react": "~18.2.60",
+    "@types/react": "~18.2.64",
     "@types/react-native": "0.73.0",
-    "@typescript-eslint/eslint-plugin": "^7.1.0",
-    "@typescript-eslint/parser": "^7.1.0",
+    "@typescript-eslint/eslint-plugin": "^7.1.1",
+    "@typescript-eslint/parser": "^7.1.1",
     "babel-eslint": "^10.1.0",
     "commitlint": "^19.0.3",
     "del-cli": "^5.1.0",
     "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-airbnb-typescript": "^18.0.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-no-relative-import-paths": "^1.5.3",
     "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react": "^7.34.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "jest": "^29.7.0",
     "prettier": "^3.2.5",
@@ -82,7 +82,7 @@
     "react-native": "^0.73.5",
     "react-native-builder-bob": "^0.23.2",
     "release-it": "^17.1.1",
-    "typescript": "^5.3.3"
+    "typescript": "^5.4.0"
   },
   "resolutions": {
     "@types/react": "17.0.21"
