+++ src/index.ts
@@ -1,0 +1,1 @@
+export { EbtBalanceFlow, EbtBalanceFlowEnvironment } from '@ebt-balance/components/EbtBalanceFlow/EbtBalanceFlow';
